composer-app: factor shared core plugins into plugin-defs.core
