+++ packages/apps/composer-app/vite.config.ts
@@ -288,7 +288,7 @@ export default defineConfig((env) => ({
       isMinimalPluginSet
         ? path.resolve(
             rootDir,
-            'packages/plugins/plugin-{assistant,attention,client,deck,graph,markdown,navtree,observability,onboarding,registry,settings,space,status-bar,theme,thread}/src/index.{ts,tsx}',
+            'packages/plugins/plugin-{assistant,attention,client,comments,deck,graph,markdown,navtree,observability,onboarding,registry,settings,simple-layout,space,spotlight,status-bar,theme,thread}/src/index.{ts,tsx}',
           )
         : path.resolve(rootDir, 'packages/plugins/*/src/index.{ts,tsx}'),
     ],
